update id types for entiteis with less than 2^32 instances expected

+++ lesson13_MigratingDB/FabricMarket_DomainModel/Models/FabricProduction/FabricType.cs
@@ -1,6 +1,6 @@
 ﻿namespace lesson11_FabricMarket_DomainModel.Models.FabricProduction
 {
-    public class FabricType : Entity<long>
+    public class FabricType : Entity<int>
     {
         public required string Name { get; set; }
         public required string Description { get; set; }
