+++ lesson13_MigratingDB/FabricMarket_DomainModel/Models/FabricProduction/ProductRequiresFabric.cs
@@ -5,7 +5,7 @@ namespace lesson11_FabricMarket_DomainModel.Models.FabricProduction
     public class ProductRequiresFabric : IEntityWithoutOwnId
     {
         public required long ProductId { get; set; }
-        public required long FabricVariantId { get; set; }
+        public required int FabricVariantId { get; set; }
         public required decimal AreaOfFabric { get; set; }
 
         public required FabricVariant FabricVariant { get; set; }
